Analyze Refined Finna Metadata

Performs basic analysis on Finna metadata, summarizing the distribution of formats, years, and authors.

Usage

analyze_metadata(metadata)

Arguments

metadata

A tibble containing refined Finna metadata.

Value

A list of tibbles with summaries of formats, years, and authors.

Examples

library(finna)
sibelius_data <- search_finna("sibelius")
#> Warning: Default limit of 100 records is being used. Specify 'limit' argument for more records.
#> Total results found: 44512
#> Data fetching completed in 2.90 seconds.
refined_data <- refine_metadata(sibelius_data)
analyze_metadata(refined_data)
#> $format_distribution
#> # A tibble: 11 × 2
#>    Formats                        n
#>    <chr>                      <int>
#>  1 Kirja, Kirja                  50
#>  2 Lehti/Artikkeli, Artikkeli    18
#>  3 Äänite, CD                    15
#>  4 Äänite, Äänilevy               6
#>  5 Video, DVD                     4
#>  6 Video, Elokuva, lyhyt          2
#>  7 Taideteos, Taideteos           1
#>  8 Taideteos, Veistos             1
#>  9 Video, Elokuva, pitkä          1
#> 10 Äänite, Musiikkitallenne       1
#> 11 Äänite, Äänikasetti            1
#> 
#> $year_distribution
#> # A tibble: 40 × 2
#>    Year      n
#>    <chr> <int>
#>  1 1948      8
#>  2 1997      8
#>  3 1945      7
#>  4 1998      6
#>  5 1968      5
#>  6 1999      5
#>  7 2003      5
#>  8 1996      4
#>  9 1949      3
#> 10 1995      3
#> # ℹ 30 more rows
#> 
#> $author_distribution
#> # A tibble: 57 × 2
#>    Author                                        n
#>    <chr>                                     <int>
#>  1 Häyrynen, Antti                              13
#>  2 Sibelius, Jean                               13
#>  3 Downes, Olin, Sjöblom, Paul, Jalas, Jussi     3
#>  4 Lampila, Hannu-Ilari                          3
#>  5 Pickenhayn, Jorge Oscar                       3
#>  6 Ringbom, Nils-Eric                            3
#>  7 Tawaststjerna, Erik                           3
#>  8 Barnett, Andrew                               2
#>  9 Gray, Cecil                                   2
#> 10 Haapaniemi, Jaakko                            2
#> # ℹ 47 more rows
#>