Fetch Finna Collection Data with Flexible Query

This function retrieves data from the Finna API and formats it as a tidy tibble.

Usage

fetch_finna_collection(
  query = NULL,
  limit = 0,
  facets = "building",
  lng = "fi",
  prettyPrint = TRUE
)

Arguments

query

The query string for filtering results. Defaults to NULL, which fetches data without a specific search term.

limit

Maximum number of results to fetch. Defaults to 0.

facets

Facet to retrieve, defaults to "building".

lng

Language for results, defaults to "fi".

prettyPrint

Logical, whether to pretty-print JSON responses.

Value

A tibble containing the fetched data with relevant fields.

Examples

fetch_finna_collection(query = "record_format:ead", limit = 0)
#> Total results found: 2062412
#> Data fetching completed in 1.37 seconds.
#> # A tibble: 3 × 4
#>   value                       translated                   count href           
#>   <chr>                       <chr>                        <int> <chr>          
#> 1 0/KA/                       Kansallisarkisto           2060481 ?lookfor=recor…
#> 2 0/KOTUS/                    Kotimaisten kielten keskus     614 ?lookfor=recor…
#> 3 0/siirtolaisuusinstituutti/ Siirtolaisuusinstituutti      1317 ?lookfor=recor…
fetch_finna_collection() # Fetches data with no specific query
#> Total results found: 18992285
#> Data fetching completed in 7.66 seconds.
#> # A tibble: 219 × 4
#>    value                          translated                   count href       
#>    <chr>                          <chr>                        <int> <chr>      
#>  1 0/3AMK/                        3AMK-kirjastot              127475 ?type=AllF…
#>  2 0/AALTO/                       Aalto-yliopisto             238151 ?type=AllF…
#>  3 0/AALTOARKISTO/                Aalto-yliopiston arkisto     64408 ?type=AllF…
#>  4 0/Aipa/                        Aipa                           122 ?type=AllF…
#>  5 0/lasipalatsi/                 Albumit auki                  2413 ?type=AllF…
#>  6 0/ALVARAALTO/                  Alvar Aalto -museo             329 ?type=AllF…
#>  7 0/Anders/                      Anders-kirjastot            268246 ?type=AllF…
#>  8 0/ARCADA/                      Arcadan kirjasto             25323 ?type=AllF…
#>  9 0/MFA/                         Arkkitehtuurimuseo           36003 ?type=AllF…
#> 10 0/Asikkalan kulttuuripalvelut/ Asikkalan kulttuuripalvelut   2718 ?type=AllF…
#> # ℹ 209 more rows