Refine Finna Metadata

The refine_metadata function cleans and standardizes Finna metadata by:

Validating Required Fields: Checks for the presence of key metadata fields and returns NULL if any are missing.

Handling Missing Values: Replaces NA values in critical fields with descriptive placeholder text (e.g., "Unknown Title").

Selecting Relevant Fields: Keeps only the following fields for streamlined analysis:

Title: The title of the resource.

Author: The creator or author of the resource.

Year: The publication or release year.

Language: The language of the resource.

Formats: The format(s) of the resource (e.g., Book, Audio).

Subjects: The subject keywords or classifications.

Library: The owning library or institution.

Series: The series or collection the resource belongs to.

Usage

refine_metadata(data)

Arguments

data

A tibble containing raw Finna metadata.

Value

A tibble with selected, cleaned metadata fields, or NULL if required fields are missing.

Examples

library(finna)
sibelius_data <- search_finna("sibelius")
#> Warning: Default limit of 100 records is being used. Specify 'limit' argument for more records.
#> Total results found: 44515
#> Data fetching completed in 0.33 seconds.
refine_metadata(sibelius_data)
#> # A tibble: 100 × 8
#>    Title                   Author Year  Language Formats Subjects Library Series
#>    <chr>                   <chr>  <chr> <chr>    <chr>   <chr>    <chr>   <chr> 
#>  1 Sibelius favourites : … Sibel… 2001  Unknown… Äänite… orkeste… Lapin … Unkno…
#>  2 Sibelius                Lampi… 1995  fin      Kirja,… Sibeliu… Anders… Unkno…
#>  3 Sibelius                Lampi… 1984  fin      Kirja,… Sibeliu… Anders… Unkno…
#>  4 Sibelius                Tawas… 1997  fin      Kirja,… Sibeliu… Rutakk… Unkno…
#>  5 Sibelius                RINGB… 1948  swe      Kirja,… Unknown… Fredri… Unkno…
#>  6 Sibelius                Ringb… 1948  swe      Kirja,… Sibeliu… Lukki-… Unkno…
#>  7 Sibelius                Ringb… 1948  swe      Kirja,… Jean Si… Svensk… Unkno…
#>  8 Sibelius                Barne… 2007  eng      Kirja,… Sibeliu… Helmet… Unkno…
#>  9 Sibelius.               Ottaw… 1968  Unknown… Kirja,… Unknown… Åbo Ak… Unkno…
#> 10 Sibelius                Sibel… 2017  zxx      Äänite… Unknown… Keski-… Unkno…
#> # ℹ 90 more rows